requirements.txt
ticktick-py
requests
bs4
python-dotenv